[RA-42] Get logging configuration from appsettings.json
