+++ bitprim.insight/Program.cs
@@ -15,7 +15,6 @@ namespace bitprim.insight
 
         public static void Main(string[] args)
         {
-            ConfigureLogging();
             try
             {
                 var config = new ConfigurationBuilder()
@@ -53,28 +52,5 @@ namespace bitprim.insight
             }
         }
 
-        private static void ConfigureLogging()
-        {
-            var timeZone = DateTimeOffset.Now.ToString("%K").Replace(CultureInfo.CurrentCulture.DateTimeFormat.TimeSeparator, "");
-            const string outputTemplate = "{" + LogPropertyNames.SOURCE_IP + "} " +
-                    "{" + LogPropertyNames.USER_ID +  "} {" + LogPropertyNames.USER_NAME  + "} " +
-                    "[{Timestamp:dd/MMM/yyyy HH:mm:ss} {" + LogPropertyNames.TIME_ZONE + "}] {Level:u3} " +
-                    "\"{" + LogPropertyNames.HTTP_METHOD + "} " +
-                    "{" + LogPropertyNames.HTTP_REQUEST_URL + "} " +
-                    "{" + LogPropertyNames.HTTP_PROTOCOL_VERSION + "}\" " +
-                    "{" + LogPropertyNames.HTTP_RESPONSE_STATUS_CODE + "} " +
-                    "{" + LogPropertyNames.HTTP_RESPONSE_LENGTH + "} " +
-                    "{Message:lj}{NewLine}{Exception}";
-            Log.Logger = new LoggerConfiguration()
-                .MinimumLevel.Debug()
-                //.MinimumLevel.Error()
-                .MinimumLevel.Override("Microsoft", LogEventLevel.Information)
-                .Enrich.FromLogContext()
-                .Enrich.WithProperty(LogPropertyNames.TIME_ZONE, timeZone)
-                .WriteTo.Console(outputTemplate: outputTemplate)
-                .WriteTo.File("log-.txt", rollingInterval: RollingInterval.Day, outputTemplate: outputTemplate)
-                .MinimumLevel.Override("Microsoft", LogEventLevel.Warning)
-                .CreateLogger();
-        }
     }
 }
