+++ bitprim.insight/appsettings.example.json
@@ -9,5 +9,34 @@
   "StartDatabaseFromScratch": true,
   "TransactionsByAddressPageSize": 10,
   "InitializeNode" : false,
-  "ForwardUrl" : "http://localhost:1234" 
+  "ForwardUrl" : "http://localhost:1234",
+  "Serilog":
+  {
+    "Using": ["Serilog.Sinks.Console", "Serilog.Sinks.File"],
+    "MinimumLevel":
+    {
+      "Default": "Debug"
+    },
+    "WriteTo":
+    [
+      {
+        "Name": "Console",
+        "Args":
+        {
+           "outputTemplate" : "{SourceIP} {UserID} {UserName} [{Timestamp:dd/MMM/yyyy HH:mm:ss} {TimeZone}] {Level:u3} {HttpMethod} {HttpRequestUrl} {HttpProtocol} {HttpResponseStatusCode} {HttpResponseLength} {Message:lj}{NewLine}{Exception}"
+        }
+      },
+      {
+        "Name": "File",
+        "Args":
+        {
+           "path": "log-.txt",
+           "rollingInterval": "Day",
+           "outputTemplate" : "{SourceIP} {UserID} {UserName} [{Timestamp:dd/MMM/yyyy HH:mm:ss} {TimeZone}] {Level:u3} {HttpMethod} {HttpRequestUrl} {HttpProtocol} {HttpResponseStatusCode} {HttpResponseLength} {Message:lj}{NewLine}{Exception}"
+        }
+      }
+    ],
+    "Enrich": ["FromLogContext"],
+    "Properties": { "Application": "Sample" }
+  }
 }
